Address PR #7467 review feedback for PMC test migration
